+++ release-notes.md
@@ -2,4 +2,4 @@
 - 支持上传自定义日程
 
 ## @thu-info/lib
-- feat: implement `saveCustomSchedule`
+- feat: implement `saveCustomSchedule` and `deleteCustomSchedule`